% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryFunctions.R
\name{Grad.BipLog.BIN}
\alias{Grad.BipLog.BIN}
\title{Necessary functions for gradients}
\usage{
Grad.BipLog.BIN(xt, par, k, lambda, ...)
}
\arguments{
\item{xt}{Input matrix.}

\item{par}{Parameters.}

\item{k}{Dimensions number.}

\item{lambda}{Penalization parameter.}

\item{...}{other arguments}
}
\value{
Output functions
}
\description{
These functions are necessary for the correct performance of the package.
}
\details{
These functions are necessary for the correct performance of the package.
}
\author{
Giovany Babativa <gbabativam@gmail.com>
}
