% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryFunctions.R
\name{Indsup.BIN}
\alias{Indsup.BIN}
\title{Supplementary idividuals}
\usage{
Indsup.BIN(xs, B, par, k, lambda, ...)
}
\arguments{
\item{xs}{Input matrix supplematary individuals.}

\item{B}{Resamples number.}

\item{par}{Parameters.}

\item{k}{Dimensions number.}

\item{lambda}{Penalization parameter.}

\item{...}{other arguments}
}
\description{
Supplementary idividuals
}
