% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryFunctions.R
\name{Indsup.GradBIN}
\alias{Indsup.GradBIN}
\title{Gradient supplementary idividuals}
\usage{
Indsup.GradBIN(xs, B, par, k, lambda, ...)
}
\arguments{
\item{xs}{Input matrix supplematary individuals.}

\item{B}{Resamples number.}

\item{par}{Parameters.}

\item{k}{Dimensions number.}

\item{lambda}{Penalization parameter.}

\item{...}{other arguments}
}
\description{
Gradient supplementary idividuals
}
