% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryFunctions.R
\name{J.BipLog.BIN}
\alias{J.BipLog.BIN}
\title{Necessary functions}
\usage{
J.BipLog.BIN(xt, par, k, lambda, ...)
}
\arguments{
\item{xt}{Input matrix.}

\item{par}{Parameters.}

\item{k}{Dimensions number.}

\item{lambda}{Penalization parameter.}

\item{...}{other arguments}

\item{xs}{Input matrix supplematary individuals.}

\item{B}{Resamples number.}
}
\value{
Output functions
}
\description{
These functions are necessary for the correct performance of the package.
}
\details{
These functions are necessary for the correct performance of the package.
}
\author{
Giovany Babativa <gbabativam@gmail.com>
}
