% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methylation.R
\docType{data}
\name{Methylation}
\alias{Methylation}
\title{DNA Methylation Binary Data}
\format{
A binary matrix with 50 rows (individuals) and 13 columns (CpG sites).
}
\source{
Publicly available methylation data used for illustrative purposes.
}
\usage{
Methylation
}
\description{
A binary matrix of DNA methylation measurements for a sample of individuals.
Each row represents an individual and each column a CpG site; a value of 1
indicates methylation and 0 indicates no methylation.
}
\examples{
data("Methylation")
dim(Methylation)
}
\keyword{datasets}
