% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryFunctions.R
\name{train_miss_pattern}
\alias{train_miss_pattern}
\title{missing pattern}
\usage{
train_miss_pattern(data, K = 7)
}
\arguments{
\item{data}{data frame}

\item{K}{number of folds}
}
\description{
missing pattern
}
