% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_LogBip.R
\name{cv_LogBip}
\alias{cv_LogBip}
\title{Cross-Validation for logistic biplot}
\usage{
cv_LogBip(
  data,
  k = 0:5,
  K = 7,
  method = "MM",
  type = NULL,
  plot = TRUE,
  maxit = NULL
)
}
\arguments{
\item{data}{Binary matrix.}

\item{k}{Dimensions to analyze. By default \code{k = 1:3}.}

\item{K}{folds. By default \code{K = 7}.}

\item{method}{Method to be used to estimate the parameters. By default \code{ method="MM"}}

\item{type}{For the conjugate-gradients method. Takes value 1 for the Fletcher–Reeves update, 2 for Polak–Ribiere and 3 for Beale–Sorenson.}

\item{plot}{draw the graph. By default \code{plot=TRUE}}

\item{maxit}{The maximum number of iterations. Defaults to 100 for the gradient methods, and 2000 for MM algorithm.}
}
\description{
This function run cross validation for logistic biplot
}
\examples{
data("Methylation")
cv_CG <- cv_LogBip(data = Methylation, k = 0:3, method = "CG", type = 1)
\dontrun{
cv_MM <- cv_LogBip(data = Methylation, k = 0:4, method = "MM", maxit = 1000)
}
}
\references{
Bro R and Kjeldahl K and Smilde AK. (2008). Cross-validation of component models: a critical look at current methods. Analytical and bioanalytical chemistry. 390(5):1241-1251

Wold S. (1978). Cross-validatory estimation of the number of components in factor and principal components models. Technometrics. 20(4):397--405.
}
\seealso{
\code{\link{plotBLB}, \link{pred_LB}, \link{fitted_LB}}
}
\author{
Giovany Babativa <gbabativam@gmail.com>
}
