% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogBip.R
\name{LogBip}
\alias{LogBip}
\title{Fitting a Binary Logistic Biplot using bootstrap methodology}
\usage{
LogBip(
  x,
  k = 2,
  L = 0,
  method = "CG",
  type = 1,
  plot = TRUE,
  maxit = NULL,
  endsegm = 0.9,
  label.ind = FALSE,
  ...
)
}
\arguments{
\item{x}{Binary matrix.}

\item{k}{Dimensions number. By default \code{k = 2}.}

\item{L}{Penalization parameter. By default \code{L = 0}.}

\item{method}{Method to be used to estimate the parameters. By default \code{ method="CG"}}

\item{type}{For the conjugate-gradients method. Takes value 1 for the Fletcher–Reeves update, 2 for Polak–Ribiere and 3 for Beale–Sorenson.}

\item{plot}{Plot the Bootstrap Logistic Biplot.}

\item{maxit}{The maximum number of iterations. Defaults to 100 for the gradient methods, and 500 without gradient.}

\item{endsegm}{The segment starts at 0.5 and ends at this value. By default \code{endsegm = 0.90}.}

\item{label.ind}{By default the row points are not labelled.}
}
\value{
Coordenates of the matrix A and B in resamples and Biplot
}
\description{
This function estimates the vector \eqn{\mu}, matrix A and matrix B using the optimization algorithm chosen by the user and applies a bootstrap methodology to determine the confidence ellipses.
}
\details{
Fitting when sup=TRUE ... whereas sup=FALSE ...
}
\examples{
data("Methylation")
set.seed(02052020)
### Conjugate Gradient with Fletcher and Reeves method
res.LB <- LogBip(x = Methylation)
### Quasi-Newton algorithm with BFGS method
res.LB <- LogBip(x = Methylation, method = "BFGS")
}
\references{
John C. Nash (2011). Unifying Optimization Algorithms to Aid Software System Users:optimx for R. Journal of Statistical Software. 43(9). 1--14.

John C. Nash (2014). On Best Practice Optimization Methods in R. Journal of Statistical Software. 60(2). 1--14.

Milan, L., & Whittaker, J. (1995). Application of the parametric bootstrap to models that incorporate a singular value decomposition. Applied Statistics, 44, 31–49.

Vicente-Villardon, J.L. and Galindo, M. Purificacion (2006), \emph{Multiple Correspondence Analysis and related Methods. Chapter: Logistic Biplots}. Chapman-Hall
}
\seealso{
\code{\link{plotBLB}, \link{performanceBLB}}
}
\author{
Giovany Babativa <gbabativam@gmail.com>
}
